requirements.txt
Django==1.6.2
django-autocomplete-light==1.4.13
django-simple-history==1.3.0
wsgiref==0.1.2
